Add support for ServerChan.
